# api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Nexo",
    "description": "API do Nexo. Este documento é o contrato: o cliente TypeScript do front é gerado a partir dele, nunca escrito à mão.",
    "version": "v1"
  },
  "paths": {
...
    "/contratos": {
      "get": {
        "tags": [
          "Contratos"
        ],
        "summary": "Lista os contratos recorrentes",
        "operationId": "ListarContratos",
        "parameters": [
          {
            "name": "busca",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "situacao",
            "in": "query",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/SituacaoContrato"
                }
              ]
            }
          },
          {
            "name": "pagina",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "tamanho",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          },
          {
            "name": "ordenarPor",
            "in": "query",
            "schema": {
              "enum": [
                "Codigo",
                "Cliente",
                "Valor",
                "Vencimento"
              ],
              "type": "string",
              "allOf": [
                {
                  "$ref": "#/components/schemas/OrdemDeContratos"
                }
              ],
              "default": 1
            }
          },
          {
            "name": "direcao",
            "in": "query",
            "schema": {
              "enum": [
                "Crescente",
                "Decrescente"
              ],
              "type": "string",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Direcao"
                }
              ],
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaDeContratos"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Contratos"
        ],
        "summary": "Cria um contrato",
        "operationId": "CriarContrato",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/DadosDeContrato"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContratoDetalhado"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaComProblemas"
                }
              }
            }
          }
        }
      }
    },
    "/contratos/{id}": {
      "get": {
        "tags": [
          "Contratos"
        ],
        "summary": "Abre um contrato",
        "operationId": "ObterContrato",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContratoDetalhado"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Contratos"
        ],
        "summary": "Altera um contrato",
        "operationId": "AlterarContrato",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/DadosDeContrato"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContratoDetalhado"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaComProblemas"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/contratos/gerar-mensalidades": {
      "post": {
        "tags": [
          "Contratos"
        ],
        "summary": "Gera as mensalidades de uma competência",
        "description": "Pode ser executado quantas vezes for preciso: o que já foi gerado é ignorado, não duplicado.",
        "operationId": "GerarMensalidades",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/PedidoDeGeracao"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoDaGeracao"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ContratoDetalhado": {
        "required": [
          "codigo",
          "descricao",
          "diaDeVencimento",
          "id",
          "inicioDaVigencia",
          "observacoes",
          "pessoaId",
          "situacao",
          "valor"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "codigo": {
            "type": "string"
          },
          "pessoaId": {
            "type": "string",
            "format": "uuid"
          },
          "descricao": {
            "type": "string"
          },
          "valor": {
            "type": "number",
            "format": "double"
          },
          "diaDeVencimento": {
            "type": "integer",
            "format": "int32"
          },
          "inicioDaVigencia": {
            "type": "string",
            "format": "date"
          },
          "fimDaVigencia": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "situacao": {
            "enum": [
              "Ativo",
              "Suspenso",
              "Encerrado"
            ],
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/SituacaoContrato"
              }
            ]
          },
          "observacoes": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ContratoNaLista": {
        "required": [
          "codigo",
          "codigoDaPessoa",
          "descricao",
          "diaDeVencimento",
          "id",
          "nomeDaPessoa",
          "pessoaId",
          "situacao",
          "valor"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "codigo": {
            "type": "string"
          },
          "pessoaId": {
            "type": "string",
            "format": "uuid"
          },
          "codigoDaPessoa": {
            "type": "string"
          },
          "nomeDaPessoa": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "valor": {
            "type": "number",
            "format": "double"
          },
          "diaDeVencimento": {
            "type": "integer",
            "format": "int32"
          },
          "situacao": {
            "enum": [
              "Ativo",
              "Suspenso",
              "Encerrado"
            ],
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/SituacaoContrato"
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "DadosDeContrato": {
        "required": [
          "diaDeVencimento",
          "inicioDaVigencia",
          "pessoaId",
          "situacao",
          "valor"
        ],
        "type": "object",
        "properties": {
          "pessoaId": {
            "type": "string",
            "format": "uuid"
          },
          "descricao": {
            "type": "string",
            "nullable": true
          },
          "valor": {
            "type": "number",
            "format": "double"
          },
          "diaDeVencimento": {
            "type": "integer",
            "format": "int32"
          },
          "inicioDaVigencia": {
            "type": "string",
            "format": "date"
          },
          "fimDaVigencia": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "situacao": {
            "enum": [
              "Ativo",
              "Suspenso",
              "Encerrado"
            ],
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/SituacaoContrato"
              }
            ]
          },
          "observacoes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Direcao": {
        "enum": [
          "Crescente",
          "Decrescente"
        ],
        "type": "string"
      },
...
      "OrdemDeContratos": {
        "enum": [
          "Codigo",
          "Cliente",
          "Valor",
          "Vencimento"
        ],
        "type": "string"
      },
...
      "PaginaDeContratos": {
        "required": [
          "itens",
          "pagina",
          "tamanho",
          "total",
          "totalMensalAtivo"
        ],
        "type": "object",
        "properties": {
          "itens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContratoNaLista"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "pagina": {
            "type": "integer",
            "format": "int32"
          },
          "tamanho": {
            "type": "integer",
            "format": "int32"
          },
          "totalMensalAtivo": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "PedidoDeGeracao": {
        "required": [
          "ano",
          "mes"
        ],
        "type": "object",
        "properties": {
          "ano": {
            "type": "integer",
            "format": "int32"
          },
          "mes": {
            "type": "integer",
            "format": "int32"
          },
          "contratoIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Problema": {
        "required": [
          "campo",
          "descricao",
          "sugestao",
          "titulo"
        ],
        "type": "object",
        "properties": {
          "campo": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "sugestao": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "RespostaComProblemas": {
        "required": [
          "problemas"
        ],
        "type": "object",
        "properties": {
          "problemas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Problema"
            }
          }
        },
        "additionalProperties": false
      },
...
      "ResultadoDaGeracao": {
        "required": [
          "foraDeVigencia",
          "geradas",
          "ignoradas",
          "recado"
        ],
        "type": "object",
        "properties": {
          "geradas": {
            "type": "integer",
            "format": "int32"
          },
          "ignoradas": {
            "type": "integer",
            "format": "int32"
          },
          "foraDeVigencia": {
            "type": "integer",
            "format": "int32"
          },
          "recado": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "SituacaoContrato": {
        "enum": [
          "Ativo",
          "Suspenso",
          "Encerrado"
        ],
        "type": "string"
      },
...
    }
  }
}